+++ li-utils/src/main/java/com/linkedin/common/urn/CorpuserUrn.java
@@ -1,5 +1,8 @@
 package com.linkedin.common.urn;
 
+import com.linkedin.data.template.Custom;
+import com.linkedin.data.template.DirectCoercer;
+import com.linkedin.data.template.TemplateOutputCastException;
 import java.net.URISyntaxException;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -43,4 +46,21 @@ public final class CorpuserUrn extends Urn {
   public static CorpuserUrn deserialize(String rawUrn) throws URISyntaxException {
     return createFromString(rawUrn);
   }
+
+  static {
+    Custom.registerCoercer(new DirectCoercer<CorpuserUrn>() {
+      public Object coerceInput(CorpuserUrn object) throws ClassCastException {
+        return object.toString();
+      }
+
+      public CorpuserUrn coerceOutput(Object object) throws TemplateOutputCastException {
+        try {
+          return CorpuserUrn.createFromString((String) object);
+        } catch (URISyntaxException e) {
+          throw new TemplateOutputCastException("Invalid URN syntax: " + e.getMessage(), e);
+        }
+      }
+    }, CorpuserUrn.class);
+  }
+
 }
