+++ li-utils/src/main/java/com/linkedin/common/urn/DataPlatformUrn.java
@@ -1,5 +1,9 @@
 package com.linkedin.common.urn;
 
+import com.linkedin.data.template.Custom;
+import com.linkedin.data.template.DirectCoercer;
+import com.linkedin.data.template.TemplateOutputCastException;
+
 import java.net.URISyntaxException;
 
 
@@ -25,5 +29,21 @@ public final class DataPlatformUrn extends Urn {
 
   public static DataPlatformUrn deserialize(String rawUrn) throws URISyntaxException {
     return createFromString(rawUrn);
+  }
+
+  static {
+    Custom.registerCoercer(new DirectCoercer<DataPlatformUrn>() {
+      public Object coerceInput(DataPlatformUrn object) throws ClassCastException {
+        return object.toString();
+      }
+
+      public DataPlatformUrn coerceOutput(Object object) throws TemplateOutputCastException {
+        try {
+          return DataPlatformUrn.createFromString((String) object);
+        } catch (URISyntaxException e) {
+          throw new TemplateOutputCastException("Invalid URN syntax: " + e.getMessage(), e);
+        }
+      }
+    }, DataPlatformUrn.class);
   }
 }
