+++ li-utils/src/main/java/com/linkedin/common/urn/DatasetUrn.java
@@ -1,6 +1,9 @@
 package com.linkedin.common.urn;
 
 import com.linkedin.common.FabricType;
+import com.linkedin.data.template.Custom;
+import com.linkedin.data.template.DirectCoercer;
+import com.linkedin.data.template.TemplateOutputCastException;
 import java.net.URISyntaxException;
 
 import static com.linkedin.common.urn.UrnUtils.*;
@@ -45,5 +48,21 @@ public final class DatasetUrn extends Urn {
 
   public static DatasetUrn deserialize(String rawUrn) throws URISyntaxException {
     return createFromString(rawUrn);
+  }
+
+  static {
+    Custom.registerCoercer(new DirectCoercer<DatasetUrn>() {
+      public Object coerceInput(DatasetUrn object) throws ClassCastException {
+        return object.toString();
+      }
+
+      public DatasetUrn coerceOutput(Object object) throws TemplateOutputCastException {
+        try {
+          return DatasetUrn.createFromString((String) object);
+        } catch (URISyntaxException e) {
+          throw new TemplateOutputCastException("Invalid URN syntax: " + e.getMessage(), e);
+        }
+      }
+    }, DatasetUrn.class);
   }
 }
