+++ li-utils/src/main/pegasus/com/linkedin/common/Urn.pdsc
@@ -4,7 +4,6 @@
   "namespace": "com.linkedin.common",
   "ref": "string",
   "java": {
-    "class": "com.linkedin.common.urn.Urn",
-    "coercerClass": "com.linkedin.common.urn.UrnCoercer"
+    "class": "com.linkedin.common.urn.Urn"
   }
 }
